reformat the logic of generating dtls key and cert
Since i make the dtls as a ffmpeg protocol, we need init it use ffurl_oepn after ICE binding. But before that, SDP need the fingerprint of cert
So i implement these gen or read function in whip rather than dtls because we need obtain fingerprint for sdp exchange.
So we need pass the string of key and cert into dtls from whip.

TODO:
Separate the management of the key and certificate into their own layer, independent from WHIP and DTLS. This is to prepare for supporting multiple DTLS implementations in the future.

Signed-off-by: Jack Lau <jacklau1222@qq.com>